+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ----
 
 ## [Unreleased] - y-m-d
+### Changed
+- Update the contribution docs with CLA
 
 ----
 > Unreleased changes must be tracked above this line.